requirements.txt
Django==1.4
PIL==1.1.7
South==0.7.4
distribute==0.6.10
sorl-thumbnail==11.12
wsgiref==0.1.2
yolk==0.4.3
django-constance (-> picklefield)
